fix: newarch

+++ packages/react-native-citizen-escposprinter/android/src/oldarch/CitizenEscposprinterSpec.kt
@@ -12,114 +12,114 @@ abstract class CitizenEscposprinterSpec internal constructor(context: ReactAppli
   // import android.hardware.usb.UsbDevice
   // abstract fun connect(connectType: Int, device: UsbDevice)
   abstract fun connect(
-      connectType: Int,
+      connectType: Double,
       address: String,
-      port: Int = -1,
-      timeout: Int = -1,
+      port: Double = -1.0,
+      timeout: Double = -1.0,
       promise: Promise
   )
   abstract fun disconnect(promise: Promise)
   abstract fun setEncoding(encoding: String, promise: Promise)
   abstract fun printerCheck(promise: Promise)
-  abstract fun status(type: Int, promise: Promise)
+  abstract fun status(type: Double, promise: Promise)
   abstract fun printText(
       data: String,
-      alignment: Int,
-      attribute: Int,
-      textSize: Int,
+      alignment: Double,
+      attribute: Double,
+      textSize: Double,
       promise: Promise
   )
   abstract fun printPaddingText(
       data: String,
-      attribute: Int,
-      textSize: Int,
-      length: Int,
-      side: Int,
+      attribute: Double,
+      textSize: Double,
+      length: Double,
+      side: Double,
       promise: Promise
   )
   abstract fun printTextLocalFont(
       data: String,
-      alignment: Int,
+      alignment: Double,
       fontType: String,
-      point: Int,
-      style: Int,
-      hRatio: Int,
-      vRatio: Int,
+      point: Double,
+      style: Double,
+      hRatio: Double,
+      vRatio: Double,
       promise: Promise
   )
   abstract fun printBitmap(
       data: String,
-      width: Int = ESCPOSConst.CMP_BM_ASIS,
-      alignment: Int = ESCPOSConst.CMP_ALIGNMENT_CENTER,
-      mode: Int = 0,
+      width: Double = ESCPOSConst.CMP_BM_ASIS.toDouble(),
+      alignment: Double = ESCPOSConst.CMP_ALIGNMENT_CENTER.toDouble(),
+      mode: Double = 0.0,
       promise: Promise
   )
   abstract fun printBarCode(
       data: String,
-      symbology: Int,
-      height: Int,
-      width: Int,
-      alignment: Int,
-      textPosition: Int,
+      symbology: Double,
+      height: Double,
+      width: Double,
+      alignment: Double,
+      textPosition: Double,
       promise: Promise
   )
   abstract fun printPDF417(
       data: String,
-      digits: Int,
-      steps: Int,
-      moduleWidth: Int,
-      stepHeight: Int,
-      ECLevel: Int,
-      alignment: Int,
+      digits: Double,
+      steps: Double,
+      moduleWidth: Double,
+      stepHeight: Double,
+      ECLevel: Double,
+      alignment: Double,
       promise: Promise
   )
   abstract fun printQRCode(
       data: String,
-      moduleSize: Int,
-      ECLevel: Int,
-      alignment: Int,
+      moduleSize: Double,
+      ECLevel: Double,
+      alignment: Double,
       promise: Promise
   )
   abstract fun printGS1DataBarStacked(
       data: String,
-      symbology: Int,
-      moduleSize: Int,
-      maxSize: Int,
-      alignment: Int,
+      symbology: Double,
+      moduleSize: Double,
+      maxSize: Double,
+      alignment: Double,
       promise: Promise
   )
-  abstract fun cutPaper(type: Int, promise: Promise)
-  abstract fun unitFeed(ufCount: Int, promise: Promise)
-  abstract fun markFeed(type: Int, promise: Promise)
-  abstract fun openDrawer(drawer: Int, pulseLen: Int, promise: Promise)
-  abstract fun transactionPrint(control: Int, promise: Promise)
-  abstract fun rotatePrint(rotation: Int, promise: Promise)
-  abstract fun pageModePrint(control: Int, promise: Promise)
+  abstract fun cutPaper(type: Double, promise: Promise)
+  abstract fun unitFeed(ufCount: Double, promise: Promise)
+  abstract fun markFeed(type: Double, promise: Promise)
+  abstract fun openDrawer(drawer: Double, pulseLen: Double, promise: Promise)
+  abstract fun transactionPrint(control: Double, promise: Promise)
+  abstract fun rotatePrint(rotation: Double, promise: Promise)
+  abstract fun pageModePrint(control: Double, promise: Promise)
   abstract fun clearPrintArea(promise: Promise)
   abstract fun clearOutput(promise: Promise)
   abstract fun printData(data: String, promise: Promise)
   abstract fun printNormal(data: String, promise: Promise)
-  abstract fun printNVBitmap(nvImageNumber: Int, promise: Promise)
-  abstract fun searchCitizenPrinter(connectType: Int, timeout: Int, promise: Promise)
-  abstract fun searchESCPOSPrinter(connectType: Int, timeout: Int, promise: Promise)
+  abstract fun printNVBitmap(nvImageNumber: Double, promise: Promise)
+  abstract fun searchCitizenPrinter(connectType: Double, timeout: Double, promise: Promise)
+  abstract fun searchESCPOSPrinter(connectType: Double, timeout: Double, promise: Promise)
   abstract fun printerCheckEx(
-      connectType: Int,
+      connectType: Double,
       address: String,
-      port: Int?,
-      timeout: Int?,
+      port: Double?,
+      timeout: Double?,
       promise: Promise
   )
   abstract fun openDrawerEx(
-      drawer: Int,
-      pulseLen: Int,
-      connectType: Int,
+      drawer: Double,
+      pulseLen: Double,
+      connectType: Double,
       address: String,
-      port: Int?,
-      timeout: Int?,
+      port: Double?,
+      timeout: Double?,
       promise: Promise
   )
-  abstract fun setPrintCompletedTimeout(timeout: Int, promise: Promise)
-  abstract fun setLog(mode: Int, path: String, maxSize: Int, promise: Promise)
+  abstract fun setPrintCompletedTimeout(timeout: Double, promise: Promise)
+  abstract fun setLog(mode: Double, path: String, maxSize: Double, promise: Promise)
   abstract fun getVersionCode(promise: Promise)
   abstract fun getVersionName(promise: Promise)
 }
